backend/requirements.txt
fastapi>=0.109.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn==0.24.0
asyncpg==0.29.0
aiosqlite==0.19.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0